+++ ShinRyuModManager/GameModel.cs
@@ -80,44 +80,141 @@ namespace ShinRyuModManager
 
                     string outputPath = Path.Combine(parlessDir.FullName, dir.Name + ".par");
                     Gibbed.Yakuza0.Pack.Program.Main(new string[] { dir.FullName }, outputPath);
-                    //Miscellaneous.GibbedPar.GibbedPacker.Pack(dir.FullName, Path.Combine(parlessDir.FullName, dir.Name + ".par"));
-
-                    /*
-                    var parameters = new ParArchiveWriterParameters
-                    {
-                        CompressorVersion = 0,
-                        OutputPath = Path.Combine(parlessDir.FullName, dir.Name + ".par"),
-                        IncludeDots = true
-                    };
-
-                    string nodeName = dir.Name;
-                    Node node = Util.ReadDirectory(dir.FullName, nodeName);
-                    node.SortChildren((x, y) => string.CompareOrdinal(x.Name.ToLowerInvariant(), y.Name.ToLowerInvariant()));
-
-                    var result = node.TransformWith<ParArchiveWriter, ParArchiveWriterParameters>(parameters);
-                    result.Dispose();
-                    */
                 }
 
+                Gibbed.Yakuza0.Pack.Program.Main(new string[] { parlessDir.FullName }, Path.Combine(parlessDir.Parent.FullName, hactDir.Name + ".par"));
+            }
+        }
+
+        public static void DoOEHActProcedure(MLO mlo)
+        {
+            bool hasHacts = false;
+
+            HashSet<string> hactDirs = new HashSet<string>();
+
+            foreach (var kv in mlo.Files)
+                if (kv.Item1.Contains("/hact/") && !kv.Item1.EndsWith(".par"))
+                {
+                    string mod = mlo.Mods[kv.Item2];
+
+                    string filePath = "mods/" + mod + "/" + kv.Item1;
+
+                    FileInfo file = new FileInfo(filePath);
+
+                    //get the folder from a path like this -> data/hact/h5000_some_hact/cmn/cmn.bin
+                    string hactDir = file.Directory.Parent.Name;
+
+                    if (file.Directory.Parent.Parent.Name == "hact")
+                    {
+                        string hactName = file.Directory.Parent.Name;
+                        string hactPath = file.Directory.Parent.FullName;
+
+                        if (hactDirs.Contains(hactPath))
+                            continue;
+
+                        hactDirs.Add(hactPath);
+                    }
+                    hasHacts = true;
+                }
+
+            if (!hasHacts)
+                return;
+
+            foreach (string hactDirPath in hactDirs)
+            {
+                DirectoryInfo hactDir = new DirectoryInfo(hactDirPath);
+                DirectoryInfo parlessDir = new DirectoryInfo(Path.Combine(GamePath.GetModsPath(), "Parless", "hact", hactDir.Name));
+
+                if (!parlessDir.Exists)
+                    parlessDir.Create();
+
+                foreach (DirectoryInfo dir in hactDir.GetDirectories())
+                {
+                    //We already repack ptc 
+                    if (dir.Name == "ptc" /*&& File.Exists(Path.Combine(hactDir.FullName, "ptc.par"))*/)
+                        continue;
+
+                    string outputFakeDir = Path.Combine(parlessDir.FullName, dir.Name);
+                   
+                    if(!Directory.Exists(outputFakeDir))
+                        Directory.CreateDirectory(outputFakeDir);
+
+                    string outputPath = Path.Combine(parlessDir.FullName, dir.Name + ".par");
+                    Gibbed.Yakuza0.Pack.Program.Main(new string[] { outputFakeDir }, outputPath);
+                    try
+                    {
+                        new DirectoryInfo(outputFakeDir).Delete(true);
+                    }
+                    catch
+                    {
+
+                    }
+                }
+
+                Gibbed.Yakuza0.Pack.Program.Main(new string[] { parlessDir.FullName }, Path.Combine(parlessDir.Parent.FullName, hactDir.Name + ".par"));
+            }
+        }
+
+        public static void DoDEHActProcedure(MLO mlo, string codename)
+        {
+            bool hasHacts = false;
+
+            HashSet<string> hactDirs = new HashSet<string>();
+
+            foreach (var kv in mlo.Files)
+                if (kv.Item1.Contains($"/hact_{codename}/") && !kv.Item1.EndsWith(".par"))
+                {
+                    string mod = mlo.Mods[kv.Item2];
+
+                    string filePath = "mods/" + mod + "/" + kv.Item1;
+
+                    FileInfo file = new FileInfo(filePath);
+
+                    //get the folder from a path like this -> data/hact_yazawa/h5000_some_hact/cmn/cmn.bin
+                    string hactDir = file.Directory.Parent.Name;
+
+                    if (file.Directory.Parent.Parent.Name == $"hact_{codename}")
+                    {
+                        string hactName = file.Directory.Parent.Name;
+                        string hactPath = file.Directory.Parent.FullName;
+
+                        if (hactDirs.Contains(hactPath))
+                            continue;
+
+                        bool isVanillaHact = GamePath.ExistsInDataAsPar(file.Directory.Parent.FullName);
+
+                        if (isVanillaHact)
+                            continue;
+
+                        hactDirs.Add(hactPath);
+                    }
+                    hasHacts = true;
+                }
+
+            if (!hasHacts)
+                return;
+
+            foreach (string hactDirPath in hactDirs)
+            {
+                DirectoryInfo hactDir = new DirectoryInfo(hactDirPath);
+                DirectoryInfo parlessDir = new DirectoryInfo(Path.Combine(GamePath.GetModsPath(), "Parless", "hact_" + codename, hactDir.Name));
+
+                if (!parlessDir.Exists)
+                    parlessDir.Create();
+
+                foreach (DirectoryInfo dir in hactDir.GetDirectories())
+                {
+                    //We already repack ptc 
+                    if (dir.Name == "ptc" && File.Exists(Path.Combine(hactDir.FullName, "ptc.par")))
+                        continue;
+
+                    string outputPath = Path.Combine(parlessDir.FullName, dir.Name + ".par");
+                    Gibbed.Yakuza0.Pack.Program.Main(new string[] { dir.FullName }, outputPath);
+                }
 
                 Gibbed.Yakuza0.Pack.Program.Main(new string[] { parlessDir.FullName }, Path.Combine(parlessDir.Parent.FullName, hactDir.Name + ".par"));
 
-
-                /*
-                var finalPar = new ParArchiveWriterParameters
-                {
-                    CompressorVersion = 0,
-                    OutputPath = Path.Combine(parlessDir.Parent.FullName, hactDir.Name + ".par"),
-                    IncludeDots = true
-                };
-
-                Node rootDir = Util.ReadDirectory(parlessDir.FullName, hactDir.Name);
-                rootDir.SortChildren((x, y) => string.CompareOrdinal(x.Name.ToLowerInvariant(), y.Name.ToLowerInvariant()));
-                rootDir.TransformWith<ParArchiveWriter, ParArchiveWriterParameters>(finalPar);
-                rootDir.Stream.Close();
-                rootDir.Dispose();
-                */
-
+                new DirectoryInfo(parlessDir.FullName).Delete(true);
             }
         }